+++ src/main/java/chess/domain/board/Board.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import chess.domain.piece.Bishop;
 import chess.domain.piece.Color;
@@ -8,9 +8,6 @@ import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
 import chess.domain.piece.Queen;
 import chess.domain.piece.Rook;
-import chess.domain.square.File;
-import chess.domain.square.Rank;
-import chess.domain.square.Square;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
