+++ src/main/java/chess/domain/board/File.java
@@ -1,4 +1,4 @@
-package chess.domain.square;
+package chess.domain.board;
 
 import java.util.Arrays;
